requirements.txt
networkx
pydantic
boto3
google-genai
python-dotenv
requests
matplotlib
scipy
pyvis
